+++ src/playgrounds/languages/markup.ts
@@ -1,27 +1,41 @@
 import * as path from "path";
 import { TextDocument } from "vscode";
 import * as config from "../../config";
+import { compileCode, getExtensions } from "./languageProvider";
 
 export const MARKUP_BASE_NAME = "index";
 
-export const MarkupLanguage = {
+const MarkupLanguage = {
   html: ".html",
   markdown: ".md",
   pug: ".pug"
 };
 
-export const MARKUP_EXTENSIONS = [
+const MARKUP_EXTENSIONS = [
   MarkupLanguage.html,
   MarkupLanguage.markdown,
   MarkupLanguage.pug
 ];
+
+export function getCandidateMarkupFilenames() {
+  return getMarkupExtensions().map(
+    (extension) => `${MARKUP_BASE_NAME}${extension}`
+  );
+}
+
+export function getMarkupExtensions() {
+  const customExtensions = getExtensions("markup");
+  return [...MARKUP_EXTENSIONS, ...customExtensions];
+}
 
 export async function getNewMarkupFilename() {
   const markupLanguage = await config.get("playgrounds.markupLanguage");
   return `${MARKUP_BASE_NAME}${MarkupLanguage[markupLanguage]}`;
 }
 
-export function getMarkupContent(document: TextDocument): string | null {
+export async function getMarkupContent(
+  document: TextDocument
+): Promise<string | null> {
   const content = document.getText();
   if (content.trim() === "") {
     return content;
@@ -36,8 +50,10 @@ export function getMarkupContent(document: TextDocument): string | null {
     } else if (extension === MarkupLanguage.markdown) {
       const markdown = require("markdown-it")();
       return markdown.render(content);
-    } else {
+    } else if (extension === MarkupLanguage.html) {
       return content;
+    } else {
+      return await compileCode("markup", extension, content);
     }
   } catch {
     return null;